+++ Innowise.Music/Model/AuthenticationResponse.cs
@@ -11,5 +11,6 @@ public class AuthenticationResponse
 {
     public string UserId { get; set; }
     public string Token { get; set; }
+    public string RefreshToken { get; set; }
     public string Email { get; set; }
 }