+++ Innowise.Music/Model/TokenRequestDto.cs
@@ -1,0 +1,7 @@
+namespace Innowise.Music.Model;
+
+public class TokenRequestDto
+{
+    public string Token { get; set; }
+    public string RefreshToken { get; set; }
+}