+++ Innowise.MusicIdentityServer/Models/User/TokenRequestDto.cs
@@ -1,0 +1,7 @@
+namespace Innowise.MusicIdentityServer.Models.User;
+
+public class TokenRequestDto
+{
+    public string Token { get; set; }
+    public string RefreshToken { get; set; }
+}